stack-check: 64-bit fixes

+++ os/sys/stack-check.h
@@ -54,6 +54,8 @@
 
 #include "contiki-conf.h"
 
+#include <stddef.h>
+
 /* Determine whether stack checking is supported depending on the plaform. */
 #ifdef PLATFORM_CONF_SUPPORTS_STACK_CHECK
 #if !PLATFORM_CONF_SUPPORTS_STACK_CHECK
@@ -104,8 +106,9 @@ void stack_check_init(void);
  *             In addition, this function can warn if the stack memory range
  *             has been completely used, but it cannot detect
  *             and warn if stack overflow has already taken place.
+ *             Returns SIZE_MAX when an error occurs.
  */
-int32_t stack_check_get_usage(void);
+size_t stack_check_get_usage(void);
 
 /**
  * \brief      Calculate the maximal permitted stack usage.
@@ -113,7 +116,7 @@ int32_t stack_check_get_usage(void);
  *             This function returns the number of bytes between the origin
  *             of the stack and the end of heap.
  */
-int32_t stack_check_get_reserved_size(void);
+size_t stack_check_get_reserved_size(void);
 
 /**
  * \brief      The origin point from which the stack grows (an optional #define)
